fix: D-day 사각 뱃지 공통 컴포넌트화 — 메인 티저 진행중 색상 통일
- DdayBadge 공통 컴포넌트 신설 (상태별: 예정 보라 / D-DAY 보라 솔리드 / 진행중 초록 / 종료 회색)
- 메인 티저가 상태 무관 보라로 하드코딩돼 뉴스탭(초록)과 달랐던 문제 해결 — 양쪽 모두 공통 컴포넌트 사용
- DS에 Badge/DdayBox 4종 등록, 시안 일정 카드에 진행중 상태 예시 반영
